Update create purchase orders. Create purchase order's items.

+++ angular/gestion-compras/src/app/products/product-list/product-list.component.ts
@@ -8,11 +8,11 @@ import { ProductServiceService } from 'src/app/services/product-service/product-
   styleUrls: ['./product-list.component.css'],
   providers: [ProductServiceService]
 })
-export class ProductListComponent implements OnInit{
+export class ProductListComponent implements OnInit {
 
-  productList: any[] =[];
+  productList: any[] = [];
 
-  constructor(public service: ProductServiceService){}
+  constructor(public service: ProductServiceService) { }
 
   ngOnInit(): void {
     this.productList = this.service.getProducts();
@@ -20,16 +20,21 @@ export class ProductListComponent implements OnInit{
   }
 
   sortByName(productA: Product, productB: Product) {
-    if(productA.name < productB.name){
+    if (productA.name < productB.name) {
       return -1;
-    } 
-    return (productA.name > productB.name) ? 1 : 0 ;
-    
+    }
+    return (productA.name > productB.name) ? 1 : 0;
+
   }
 
-  delete(product: Product){
-    this.service.deleteProduct(product);
-    this.productList = this.service.getProducts();
+  delete(product: Product) {
+    let confirmacion = confirm(`¿Desea eliminar el product ${product.name}?`)
+    if (confirmacion) {
+      this.service.deleteProduct(product);
+      this.productList = this.service.getProducts();
+
+    }
   }
+
 
 }
